requirement.txt
pypinyin
openai-whisper
moviepy
libsndfile
spleeter